# backend/docs/swagger.yaml
basePath: /
definitions:
  api_rest.FolderInitFile:
    properties:
      mime:
        type: string
      path:
        type: string
      sha256:
        type: string
      size:
        type: integer
    type: object
  api_rest.FolderInitRequest:
    properties:
      files:
        items:
          $ref: '#/definitions/api_rest.FolderInitFile'
        type: array
      idempotencyKey:
        type: string
      parentId:
        type: string
      rootName:
        type: string
    required:
    - files
    - rootName
    type: object
  api_rest.FolderInitResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/uploads.FolderInitFileResponse'
        type: array
      folders:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      rootFolderId:
        type: string
      uploadId:
        type: string
    type: object
  api_rest.abortRequest:
    properties:
      sessionId:
        type: string
    required:
    - sessionId
    type: object
  api_rest.abortResponse:
    properties:
      message:
        example: aborted
        type: string
    type: object
  api_rest.completeRequest:
    properties:
      clientSha256:
        type: string
      folderId:
        type: string
      sessionId:
        type: string
    type: object
...
  api_rest.createSessionRequest:
    properties:
      clientSha256:
        type: string
      declaredMime:
        type: string
      filename:
        type: string
      originalSize:
        type: integer
    type: object
  api_rest.createSessionResponse:
    properties:
      existingFileId:
        type: string
      sessionId:
        type: string
      skipUpload:
        type: boolean
      tempBlobKey:
        type: string
      uploadUrl:
        type: string
      userFileId:
        type: string
    type: object
...
  server.GraphQLRequest:
    properties:
      operationName:
        type: string
      query:
        type: string
      variables:
        additionalProperties: true
        type: object
    type: object
  server.GraphQLResponse:
    properties:
      data: {}
      errors:
        items:
          additionalProperties: true
          type: object
        type: array
    type: object
  server.WSMessage:
    properties:
      downloadCount:
        type: integer
      fileId:
        type: string
      ts:
        type: integer
    type: object
  uploads.FolderInitFileResponse:
    properties:
      deduped:
        type: boolean
      path:
        type: string
      sessionId:
        type: string
      targetFolderId:
        type: string
      tempBlobKey:
        type: string
      uploadUrl:
        type: string
      userFileId:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://example.com/support
  description: REST, GraphQL, and WebSocket endpoints for the file service.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BalkanID File Service API
  version: "1.0"
paths:
...
  /api/v1/graphql:
    post:
      consumes:
      - application/json
      description: Send GraphQL queries and mutations
      parameters:
      - description: GraphQL request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.GraphQLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.GraphQLResponse'
      security:
      - BearerAuth: []
      summary: GraphQL endpoint
      tags:
      - graphql
...
  /api/v1/uploads/abort:
    post:
      consumes:
      - application/json
      description: Cancel an ongoing upload session and clean up any partial uploads
      parameters:
      - description: Upload abort request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.abortRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Upload aborted successfully
          schema:
            $ref: '#/definitions/api_rest.abortResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Abort file upload
      tags:
      - uploads
  /api/v1/uploads/complete:
    post:
      consumes:
      - application/json
      description: Complete a multipart upload session and finalize the file
      parameters:
      - description: Upload completion request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.completeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Upload completed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Quota exceeded
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete file upload
      tags:
      - uploads
  /api/v1/uploads/folder/init:
    post:
      consumes:
      - application/json
      description: Initialize upload session for a folder structure with multiple
        files
      parameters:
      - description: Folder initialization request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.FolderInitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Upload session initialized
          schema:
            $ref: '#/definitions/api_rest.FolderInitResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Quota exceeded
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Initialize folder upload
      tags:
      - uploads
  /api/v1/uploads/session:
    post:
      consumes:
      - application/json
      description: Create a new file upload session with presigned URLs for multipart
        upload
      parameters:
      - description: Upload session request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.createSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: File already exists (skip upload)
          schema:
            $ref: '#/definitions/api_rest.createSessionResponse'
        "201":
          description: Upload session created
          schema:
            $ref: '#/definitions/api_rest.createSessionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Quota exceeded
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create upload session
      tags:
      - uploads
...
  /api/v1/ws/downloads:
    get:
      description: Connect via WebSocket to receive download count updates. Optional
        query param fileId to filter.
      parameters:
      - description: Filter by file ID
        in: query
        name: fileId
        type: string
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "200":
          description: Example server message (for documentation only)
          schema:
            $ref: '#/definitions/server.WSMessage'
      summary: WebSocket for download updates
      tags:
      - websocket
securityDefinitions:
  BearerAuth:
    description: 'Provide a Bearer token: "Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"